requirements.txt
Django>=4
djangorestframework>=3
beautifulsoup4>=4
pytz